[63] Click coordinates converted=>Spell animation works. Also some cleanup

+++ client/Math.cpp
@@ -20,7 +20,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 namespace math
 {
-    float GetAngle(sf::Vector2f Origin, sf::Vector2i Target)
+    float GetAngle(sf::Vector2f Origin, sf::Vector2f Target)
     {
         // Click position should represent center of sprite, NOT the upper-left corner
         Target.x -= TILE_SIZE / 2;
